+++ Tests/Unit/MSFT_xSQLServerMaxDop.Tests.ps1
@@ -13,6 +13,9 @@ if ( (-not (Test-Path -Path (Join-Path -Path $script:moduleRoot -ChildPath 'DSCR
 
 Import-Module (Join-Path -Path $script:moduleRoot -ChildPath 'DSCResource.Tests\TestHelper.psm1') -Force
 
+# Loading mocked classes
+Add-Type -Path ( Join-Path -Path ( Join-Path -Path $PSScriptRoot -ChildPath Stubs ) -ChildPath SMO.cs )
+
 $TestEnvironment = Initialize-TestEnvironment `
     -DSCModuleName $script:DSCModuleName `
     -DSCResourceName $script:DSCResourceName `
@@ -40,6 +43,7 @@ try
         $mockInvalidOperationForAlterMethod = $false
         $mockNumberOfLogicalProcessors      = 4
         $mockNumberOfCores                  = 4
+        $mockProcessOnlyOnActiveNode        = $true
 
         # Default parameters that are used for the It-blocks
         $mockDefaultParameters = @{
@@ -52,9 +56,9 @@ try
         $mockConnectSQL = {
             return @(
                 (
-                    New-Object Object |
-                        Add-Member -MemberType NoteProperty -Name InstanceName -Value $mockSQLServerInstanceName -PassThru |
-                        Add-Member -MemberType NoteProperty -Name ComputerNamePhysicalNetBIOS -Value $mockSQLServerName -PassThru |
+                    New-Object Object -TypeName Microsoft.SqlServer.Management.Smo.Server |
+                        Add-Member -MemberType NoteProperty -Name InstanceName -Value $mockSQLServerInstanceName -PassThru -Force |
+                        Add-Member -MemberType NoteProperty -Name ComputerNamePhysicalNetBIOS -Value $mockSQLServerName -PassThru -Force |
                         Add-Member -MemberType ScriptProperty -Name Configuration -Value {
                             return @( ( New-Object Object |
                                 Add-Member -MemberType ScriptProperty -Name MaxDegreeOfParallelism -Value {
@@ -96,6 +100,7 @@ try
 
         Describe "MSFT_xSQLServerMaxDop\Get-TargetResource" -Tag 'Get'{
             Mock -CommandName Connect-SQL -MockWith $mockConnectSQL -Verifiable
+            Mock -CommandName Test-ActiveNode -MockWith { return $mockProcessOnlyOnActiveNode } -Verifiable
 
             Context 'When the system is either in the desired state or not in the desired state' {
                 $testParameters = $mockDefaultParameters
@@ -114,6 +119,10 @@ try
                 It 'Should call the mock function Connect-SQL' {
                     Assert-MockCalled Connect-SQL -Exactly -Times 1 -Scope Context
                 }
+
+                It 'Should call the mock function Test-ActiveNode' {
+                    Assert-MockCalled Test-ActiveNode -Exactly -Times 1 -Scope Context
+                }
             }
 
             Assert-VerifiableMock
@@ -122,6 +131,10 @@ try
         Describe "MSFT_xSQLServerMaxDop\Test-TargetResource" -Tag 'Test'{
             BeforeEach {
                 Mock -CommandName Connect-SQL -MockWith $mockConnectSQL -Verifiable
+
+                Mock -CommandName Test-ActiveNode -MockWith {
+                    $mockProcessOnlyOnActiveNode
+                } -Verifiable
 
                 Mock -CommandName Get-CimInstance -MockWith $mockCimInstance_Win32Processor -ParameterFilter {
                     $ClassName -eq 'Win32_Processor'
@@ -283,6 +296,31 @@ try
                 It 'Should return the state as false when desired MaxDop is the wrong value' {
                     $result = Test-TargetResource @testParameters
                     $result | Should -Be $false
+                }
+
+                It 'Should call the mock function Connect-SQL' {
+                    Assert-MockCalled Connect-SQL -Exactly -Times 1 -Scope Context
+                }
+            }
+
+            Context 'When the ProcessOnlyOnActiveNode parameter is passed' {
+                AfterAll {
+                    $mockProcessOnlyOnActiveNode = $true
+                }
+
+                BeforeAll {
+                    $testParameters = $mockDefaultParameters
+                    $testParameters += @{
+                        Ensure = 'Absent'
+                        ProcessOnlyOnActiveNode = $true
+                    }
+
+                    $mockProcessOnlyOnActiveNode = $false
+                }
+
+                It 'Should return $true when ProcessOnlyOnActiveNode is "$true" and the current node is not actively hosting the instance' {
+                    $result = Test-TargetResource @testParameters
+                    $result | Should -Be $true
                 }
 
                 It 'Should call the mock function Connect-SQL' {